Fixed issue with Murphy plugin using wronge type for certain Properties.
Murphy tries to create a basic type even if the type is already registered.
This wouldn't typically be a problem, except that certain types are enums
which murphy doesn't know about.

The solution is to use VehicleProperty::getPropertyTypeForPropertyName if the
property has already been registered by another plugin or by AMB core itself.